requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
APScheduler==3.10.1
python-dotenv==1.0.0
influxdb-client
python-dateutil
